post-producer/requirements.txt
python-dotenv==1.0.1
openai==1.13.3
pillow==10.2.0
pytest==8.3.2